refactor: command, query 모델 패키지 구조 개선
