+++ src/main/kotlin/com/server/dpmcore/member/member/domain/port/inbound/MemberQueryUseCase.kt
@@ -1,7 +1,7 @@
 package com.server.dpmcore.member.member.domain.port.inbound
 
 import com.server.dpmcore.member.member.domain.model.MemberId
-import com.server.dpmcore.member.member.domain.port.inbound.query.MemberNameAuthorityQueryModel
+import com.server.dpmcore.member.member.domain.model.query.MemberNameAuthorityQueryModel
 
 interface MemberQueryUseCase {
     /**
